# Detector speedups and Divizion by zero fix in cli