@Breakthrough 2018-08-06
Just fixed the divide by zero error in parallel, I'll fix the merge conflicts, no worries.  Before merging the rest, however, I would like to look a little bit into the detector sensitivity changes before merging this.  The equation for computing the threshold looks almost identical, do you know why they are different?  (No worries if not, I can also look into it when I get some time)