@Breakthrough 2019-05-20
Hi @bubalazi;

Would you be willing to split this pull request into two?  One to address the divide by zero, and the other to address the speedup?

Thank you!