@Breakthrough 2019-07-21
Hi @bubalazi;

My apologies, I will need to reject this PR for now.  This PR needs to be rebased and I'm not 100% sure which changes overlap with the existing changes in master.

Furthermore, due to accuracy differences noted, the unit tests will also need updating.  Unfortunately those issues need to be resolved before this PR can be pulled into master.  Sorry about that, I hope you understand.  If you still feel the speed gains are worth it, and want these fixes implemented in a future version of PySceneDetect, you can:

 - Rebase this PR and 
 - Submit a new PR
 - Submit a new issue with suggested fix (reference this PR)

Sorry again about the delay in getting to this, have been busy with work for most of this year, and am finally getting around to cleaning up everything since the v0.5 release.

Thank you so much for your patience, I really do appreciate your help w/ the project!