@Breakthrough 2019-07-22
Divide by zero is fixed by #89 (instead of modifying the list length, instead the whole video is returned as a single scene).